+++ src/knockback-sample-transitions-jquery/module-loader.js
@@ -1,0 +1,18 @@
+/*
+  knockback-sample-transitions-jquery.js 0.1.1
+  (c) 2011, 2012 Kevin Malakoff - http://kmalakoff.github.com/knockback/
+  License: MIT (http://www.opensource.org/licenses/mit-license.php)
+  Dependencies: Knockout.js, Backbone.js, and Underscore.js.
+*/
+(function() {
+  return (function(factory) {
+    // AMD
+    if (typeof define === 'function' && define.amd) {
+      return define('knockback-sample-transitions-jquery', factory);
+    }
+    // CommonJS/NodeJS or No Loader
+    else {
+      return factory.call(this);
+    }
+  })(function() {'__REPLACE__'; return kb;});
+}).call(this);