+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@ Please refer to the following release notes when upgrading your version of Knock
 ## 0.1.1
 
 * updated kb.loadUrl and added kb.loadUrlFn so they can be called directly from an HTML View
+* renamed knockback-sample-transitions to knockback-sample-transitions-jquery
 
 ## 0.1.0